--- resources/__init__.py
@@ -1,24 +1,0 @@
-# IMPORT DEPENDENCIES
-import logging
-import colorlog
-
-### INITIALIZE LOGGING ###
-def setup_logging(logging_level=logging.INFO):
-    log = logging.getLogger(__file__)
-    log.setLevel(logging_level)
-    format_str = '%(bold_black)s%(asctime)s %(log_color)s%(levelname)-8s %(purple)s%(filename)s%(reset)s %(message)s'
-    date_format = '%Y-%m-%d %H:%M:%S'
-    cformat = f'{format_str}'
-    colors={
-            'DEBUG': 'green',
-            'INFO': 'cyan',
-            'WARNING': 'yellow',
-            'ERROR': 'red',
-            'CRITICAL': 'red,bg_white',
-        }
-    formatter = colorlog.ColoredFormatter(cformat, date_format, log_colors=colors)
-    stream_handler = logging.StreamHandler()
-    stream_handler.setFormatter(formatter)
-    log.addHandler(stream_handler)
-    return log
-log = setup_logging(logging_level=logging.DEBUG)
